+++ Demo/Demo/NAAnimatedDemoViewController.m
@@ -22,7 +22,8 @@
     mapView.backgroundColor  = [UIColor colorWithRed:0.000f green:0.475f blue:0.761f alpha:1.000f];
     mapView.autoresizingMask = UIViewAutoresizingFlexibleWidth | UIViewAutoresizingFlexibleHeight;
 
-    [mapView displayMap:[UIImage imageNamed:@"Maps/australia"]];
+    NSString *australia = [[[NSBundle mainBundle] bundlePath] stringByAppendingString:@"/Maps/australia.png"];
+    [mapView displayMap:[UIImage imageWithContentsOfFile:australia]];
 
     mapView.minimumZoomScale = 0.5f;
     mapView.maximumZoomScale = 1.5f;