+++ Demo/Demo/NALoadViaNIBDemoViewController.m
@@ -15,7 +15,9 @@
 {
     [super viewDidLoad];
 
-    [self.mapView displayMap:[UIImage imageNamed:@"Maps/australia"]];
+    NSString *australia = [[[NSBundle mainBundle] bundlePath] stringByAppendingString:@"/Maps/australia.png"];
+    UIImage *image = [UIImage imageWithContentsOfFile:australia];
+    [self.mapView displayMap:image];
 
     self.mapView.backgroundColor = [UIColor colorWithRed:0.000f green:0.475f blue:0.761f alpha:1.000f];