+++ Demo/Demo/main.m
@@ -3,7 +3,7 @@
 //  Demo
 //
 //  Created by Neil Ang on 6/05/12.
-//  Copyright (c) 2012 __MyCompanyName__. All rights reserved.
+//  Copyright (c) 2010-14 neilang.com. All rights reserved.
 //
 
 #import <UIKit/UIKit.h>