+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 #### Next
 
 * [#31](https://github.com/neilang/NAMapKit/issues/31) - Fix: `NAPinAnnotation` and `NAPinAnnotationView` retain cycle - [@dblock](https://github.com/dblock).
+* [#34](https://github.com/neilang/NAMapKit/issues/34) - Fix: `NAPinAnnotationCallOutView` uses `imageNamed` which retains memory - [@dblock](https://github.com/dblock).
 
 #### [3.1](https://github.com/neilang/NAMapKit/tree/v3.1) (5/1/2014)